I might add this information on the wiki
This is revision 277 from my local repository and the first revision online. Use 1.6 compiler.

Note, that the project actually contains to projects:
- bolscript: The tabla notation project. The bolscript source folders are bolscript_src and bolscript_tests and they are independent of the other group called gat
Main class is bolscript/Master[Mac]

- gat: a tabla composition project I started 2006 in my bachelor report and haven't updated much since then.
I want to keep it in here, so it keeps working when refactorings in bolscript are done. (I moved all components which are used in bolscript to the bolscript folders, so gat depends on bolscript)
Main class is managing/Manager.

Both projects need junit 4. 
Bolscript needs references to iText (needed for pdf writing) and gat needs swixml. 
I am not sure if .classpath,.project and eclipse settings should be included in the svn, so I left them unversioned, and please keep it that way or tell me why otherwise would be smarter.

Running bolscript:
start bolscript/Master[Mac]. Copy the resources/tablafolder_default somewhere on your system, and choose it as tabla folder when first running bolscript. Running bolscript generates a settings.xml. Do not add it to svn please.